GSF.TimeSeries & GSF.PhasorProtocols: Fixed a bug where configuration cache backups were still being created after setting ConfigurationBackups to 0 in the config file.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/gsf/Main;C90279

+++ Source/Libraries/GSF.PhasorProtocols/Anonymous/ConfigurationFrame.cs
@@ -217,15 +217,18 @@ namespace GSF.PhasorProtocols.Anonymous
 
                 try
                 {
-                    // Back up current configuration file, if any
-                    if (File.Exists(configurationCacheFileName))
+                    if (ConfigurationBackups > 0)
                     {
-                        string backupConfigFile = configurationCacheFileName + ".backup";
+                        // Back up current configuration file, if any
+                        if (File.Exists(configurationCacheFileName))
+                        {
+                            string backupConfigFile = configurationCacheFileName + ".backup";
 
-                        if (File.Exists(backupConfigFile))
-                            File.Delete(backupConfigFile);
+                            if (File.Exists(backupConfigFile))
+                                File.Delete(backupConfigFile);
 
-                        File.Move(configurationCacheFileName, backupConfigFile);
+                            File.Move(configurationCacheFileName, backupConfigFile);
+                        }
                     }
                 }
                 catch (Exception ex)
